fix: pre-audit security fixes across multiple crates
Fixes from internal security audit before external audit engagement:

- consensus: eliminate AtomicU64 call-order coupling in L1 message validation,
  split into independent stateless checks (validate_l1_messages_in_block +
  header monotonicity check)
- consensus: add timestamp validation for pre-Emerald (strict >) vs
  post-Emerald (>=)
- revm/handler: add disable_fee_charge check in token fee path for eth_call
  simulation correctness
- revm/handler: handle missing cached_token_fee_info in reimburse_caller
  when fee charge was disabled
- revm/handler: use checked_sub in transfer_erc20_with_slot for from balance
- revm/l1block: cap L1 data fee to u64::MAX for circuit compatibility
- engine-api: fix ValidateL2Block error semantics (internal vs invalid)
- engine-api: fix FeeVault coinbase override
- rpc: use TxMorph::validate() instead of validate_version() for complete
  validation at RPC layer
- primitives: minor TxMorph validation improvements
- txpool: improve error types for MorphTx validation

+++ crates/revm/src/tx.rs
@@ -140,7 +140,7 @@ impl MorphTxEnv {
         Some(TxMorph {
             chain_id: self.chain_id().unwrap_or(fallback_chain_id),
             nonce: self.inner.nonce,
-            gas_limit: self.gas_limit() as u128,
+            gas_limit: self.gas_limit(),
             max_fee_per_gas: self.max_fee_per_gas(),
             max_priority_fee_per_gas: self.max_priority_fee_per_gas().unwrap_or_default(),
             to: self.kind(),
